Add help text to the Build Trigger entry.
This is very much needed. I didn't find any documentation of the job parameters
anywhere, so I had to resort to grepping the code to figure it out, which is
obviously not an ideal user experience.

`git grep 'params.put'` was used to determine where the plugin assembles the
job parameters. This returned `ActiveMqMessagingWorker` and
`FedMsgMessagingWorker`, so I've documented those two as best I can.

`help-trigger.html` is added to `webapp/`, and linked to by
`CIBuildTriggerDescriptor#getHelpFile()`.