Use durable queue consumers when subscribing jobs in ActiveMq
For some reason bee4845, the commit that added queue support, didn't mirror the
durability of topic subscriptions to queue consumers. I suspect this is the
cause of sporadic dropped messages I've seen when the provider is configured to
use queues.

`ActiveMqMessagingWorker#subscribe` is the smallest and most important change.

I applied a similar fix to `ActiveMqMessagingWorker#waitForMessage` and
`ActiveMqMessageWatcher#watch`, but actual durability in those cases isn't
possible without making them to retry on a timeout/error. I suspect that
sporadic dropped messages will still be a problem when a job waits for a
message while executing.